+++ app/src/main/java/zechs/drive/stream/di/AppModule.kt
@@ -9,6 +9,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import zechs.drive.stream.data.local.AccountsDao
 import zechs.drive.stream.utils.AppSettings
+import zechs.drive.stream.utils.FirstRunProfileMigrator
 import zechs.drive.stream.utils.SessionManager
 import javax.inject.Singleton
 
@@ -36,5 +37,16 @@ object AppModule {
     fun provideThemeDataStore(
         @ApplicationContext appContext: Context
     ): AppSettings = AppSettings(appContext)
+
+    @Provides
+    @Singleton
+    fun provideFirstRunProfileMigrator(
+        @ApplicationContext appContext: Context,
+        gson: Gson,
+        sessionManager: SessionManager,
+        accountsManager: AccountsDao,
+    ): FirstRunProfileMigrator {
+        return FirstRunProfileMigrator(appContext, gson, sessionManager, accountsManager)
+    }
 
 }
